requirements-jnpl.txt
click-odoo==1.4.2
python-barcode==0.11.0
z3c.rml==3.10.0
